% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_lsoa}
\alias{add_lsoa}
\title{Add LSOA}
\usage{
add_lsoa(data, nspl)
}
\arguments{
\item{data}{Data frame.}

\item{nspl}{NSPL lookup data frame.}
}
\value{
Data frame with added lsoa column.
}
\description{
Add LSOA
}
