% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dths_imd.R
\name{add_most_least_deprived_label}
\alias{add_most_least_deprived_label}
\title{Add most or least deprived label onto IMD factors}
\usage{
add_most_least_deprived_label(data)
}
\arguments{
\item{data}{Formatted data frame, with \verb{IMD quantile group} column.}
}
\value{
Data frame with \verb{IMD quantile group} edited to include most/least
deprived tags where applicable.
}
\description{
Add most or least deprived label onto IMD factors
}
