% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{add_zero_counts}
\alias{add_zero_counts}
\title{Add zero counts to age, sex, area, place breakdowns}
\usage{
add_zero_counts(data)
}
\arguments{
\item{data}{Data frame with following columns:
\code{week}, \code{area}, \code{sex}, \code{agegrp_5yr}, \code{place_of_death}}
}
\value{
Data frame with zero deaths for missing combinations.
}
\description{
Adds missing combinations for function creating death counts
with week, area, sex, age and place of death breakdowns. Note that not all
combinations are included, so this completes the function for the following:
week x area x age x sex AND
week x area x place of death
}
\details{
NOTE: This does \emph{NOT} work for combinations of variables that do not
exist at all (e.g. there are 0 deaths in age group 25 to 29). It only works
if there is at least one of the combinations. It must be used with care and
preferably following factors.
}
