% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{add_zero_counts_delays}
\alias{add_zero_counts_delays}
\title{Add zero counts to delay data frame.}
\usage{
add_zero_counts_delays(df)
}
\arguments{
\item{df}{Data frame of summary of delays by certification type, place of
death and area.}
}
\value{
Data frame with 0 values added where there is no occurrences of the
combination.
}
\description{
Add zero counts to delay data frame.
}
\details{
\code{cert_type}, \code{place_of_death} and \code{area} columns must be factors.
Also adds NA to the delay measure cols where there are 0 deaths.
}
