% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{add_zero_counts_no_pod}
\alias{add_zero_counts_no_pod}
\title{Add zero counts to age, sex, area, place breakdowns}
\usage{
add_zero_counts_no_pod(data)
}
\arguments{
\item{data}{Data frame with following columns:
week, area, sex, agegrp_excess}
}
\value{
Data frame with zero deaths for missing combinations.
}
\description{
Adds missing combinations for function creating death counts
with week, area, sex, age and place of death breakdowns. Note that not all
combinations are included, so this completes the function for the following:
week x area x age x sex AND
week x area
}
