% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{add_zero_counts_uk}
\alias{add_zero_counts_uk}
\title{Add zero counts of death to UK country, age, sex breakdown}
\usage{
add_zero_counts_uk(data, age_group = "agegrp_wide")
}
\arguments{
\item{data}{Data frame with following columns:
\code{year}, \code{week}, \code{country}, \code{sex} and \code{deaths}.}

\item{age_group}{string of age group, default is "agegrp_wide"}
}
\value{
Data frame with zero deaths for missing combinations.
}
\description{
Adds missing combinations for function creating death counts
with year, week, country, age and sex.
}
\details{
This will also use the week number to complete any missing week
ending dates.
}
