% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{adjust_asmr_ci_yearly}
\alias{adjust_asmr_ci_yearly}
\title{Convert weekly age-standardised mortality rate (ASMR) and confidence limits
to yearly rate}
\usage{
adjust_asmr_ci_yearly(data)
}
\arguments{
\item{data}{Data frame with columns for \code{age_standardised_rate}, \code{ci_upper},
\code{ci_lower}.}
}
\value{
Data frame with the columns adjusted from weekly rates to annual
rates.
}
\description{
Convert weekly age-standardised mortality rate (ASMR) and confidence limits
to yearly rate
}
