% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_dths.R
\name{average_5y}
\alias{average_5y}
\title{Calculate five year average}
\usage{
average_5y(df, groupings = c())
}
\arguments{
\item{df}{data frame containing data for five years by year, week, country
and desired groupings}

\item{groupings}{Character vector with additional column names to group by.}
}
\value{
data frame with five year averages by week, country and groupings
}
\description{
Calculate five year average
}
