% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dths_imd.R
\name{bind_imd_values}
\alias{bind_imd_values}
\title{Bind IMD values onto table}
\usage{
bind_imd_values(table_data, deaths_data, current_week, current_year, imd_list)
}
\arguments{
\item{table_data}{Data frame of deaths by area, sex, age group and place of
death.}

\item{deaths_data}{Deaths data frame.}

\item{current_week}{Week for analysis as numeric or string e.g. 1.}

\item{current_year}{Year for analysis as numeric e.g. 2024.}

\item{imd_list}{list containing items england and wales, the imd lookups}
}
\value{
Data frame.
}
\description{
Bind IMD values onto table
}
