% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_date_one_week_before}
\alias{calc_date_one_week_before}
\title{Calculates the date one week before a given date}
\usage{
calc_date_one_week_before(date_as_string)
}
\arguments{
\item{date_as_string}{string of 8 digits e.g. "20240131"}
}
\value{
a string of the date one week before the input string
}
\description{
Calculates the date one week before a given date
}
\details{
The date needs to be a string in format yyyymmdd. The returned
string will be in the same format
}
