% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{calc_dths_cause_dashboard}
\alias{calc_dths_cause_dashboard}
\title{Creates table of deaths by cause for evergreen dashboard}
\usage{
calc_dths_cause_dashboard(ew_deaths)
}
\arguments{
\item{ew_deaths}{Processed EW weekly deaths data frame, containing columns
region, cause, dor_year, dor_week, and week_ending.}
}
\value{
Data frame with columns registered deaths, by country and cause.
}
\description{
Modifies existing deaths data frame "ew_deaths"
to evergreen dashboard specifications. Filters and pivots to create columns
for registered deaths, by country and cause.
}
