% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{calc_excess_deaths}
\alias{calc_excess_deaths}
\title{Creates table of excess deaths}
\usage{
calc_excess_deaths(data, excess_death_data, year, week)
}
\arguments{
\item{data}{Data frame with columns week, area, sex, agegrp_excess, deaths,
week_ending}

\item{excess_death_data}{dataframe with columns dor_year, dor_week, area,
sex, deaths, expected_deaths}

\item{year}{numeric year for analysis (YYYY)}

\item{week}{numeric week for analysis}
}
\value{
Data frame with columns as data dataframe and expected_deaths
}
\description{
Creates excess deaths table by creating counts, joining with
expected deaths data and adding percentage change
}
