% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{calc_excess_deaths_uk}
\alias{calc_excess_deaths_uk}
\title{Creates table of excess deaths for all countries of the UK}
\usage{
calc_excess_deaths_uk(data, excess_death_data, config)
}
\arguments{
\item{data}{Data frame with columns week, area, sex, agegrp_excess, deaths,
week_ending}

\item{excess_death_data}{frame with columns dor_year, dor_week, area, sex,
deaths, expected_deaths}

\item{config}{list containing week (numeric week for analysis) and scope
(string of "UK" or "EW")}
}
\value{
Data frame with columns as data dataframe and expected_deaths
}
\description{
Creates excess deaths table by creating counts, joining with
expected deaths data and adding percentage change
}
