% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{calc_exp_deaths_dashboard}
\alias{calc_exp_deaths_dashboard}
\title{Creates table of expected deaths for evergreen dashboard}
\usage{
calc_exp_deaths_dashboard(exp_deaths_sql, ew_deaths, cfg)
}
\arguments{
\item{exp_deaths_sql}{Data frame containing expected death data including
area, age, sex.}

\item{ew_deaths}{Processed EW weekly deaths data frame, containing columns
region, cause, dor_year, dor_week, and week_ending.}

\item{cfg}{Config containing \code{year} and \code{week}}
}
\value{
Data frame with columns expected_deaths.
}
\description{
Modifies existing expected deaths data frame "exp_deaths_sql"
to evergreen dashboard specifications. Filters, groups and pivots to create
columns for expected deaths by country.
}
