% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_dths.R
\name{calc_p_change}
\alias{calc_p_change}
\title{Calculate percentage change for a value}
\usage{
calc_p_change(current, reference, dp = 1)
}
\arguments{
\item{current}{Current value}

\item{reference}{Refence Value}

\item{dp}{rounds to dp decimal points, default 1 decimal point}
}
\value{
Percentage change rounded to dp decimal points
}
\description{
Rounds to dp decimal points
}
