% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{calc_phe_uk_asmr}
\alias{calc_phe_uk_asmr}
\title{Calculate Age-Standardised Mortality Rates and Confidence Limits}
\usage{
calc_phe_uk_asmr(ew_df, scot_df, ni_df, config, esp, weekly_pop)
}
\arguments{
\item{ew_df}{Processed England and Wales dataframe.}

\item{scot_df}{Processed Scotland dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}

\item{config}{list containing year (numeric in YYYY format), week (the
analysis week as numeric e.g. 22) and scope (string of "UK" or "EW").}

\item{esp}{European Standardised Population (ESP) lookup data frame.}

\item{weekly_pop}{Weekly population estimates dataframe.}
}
\value{
Dataframe with PHE ASMR and CI calculated.
}
\description{
Uses the \code{PHEindicatormethods} package created by Department of
Health and Social Care to calculate the Age-Standardised Mortality Rates and
confidence intervals.
}
