% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{calc_registration_delays}
\alias{calc_registration_delays}
\title{Calculate registration delays for death data}
\usage{
calc_registration_delays(data)
}
\arguments{
\item{data}{Data frame of all year death data.}
}
\value{
Data frame with additional delay column.
}
\description{
Converts dor and reg_stat_dor to dates and
adds new column of the delay in days, throws an error if registration date
is before the occurrence date.
}
