% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_start_date_for_year}
\alias{calc_start_date_for_year}
\title{Calculates the start date for a year}
\usage{
calc_start_date_for_year(year, is_scot = FALSE)
}
\arguments{
\item{year}{Integer year in YYYY format.}

\item{is_scot}{logical, default is FALSE}
}
\value{
Numeric date of start date to use for the year in format YYYYMMDD
}
\description{
Uses the ISO format to calculate the start of the year for this
analysis from the defined week 1 (W01-1). As we want the Thursday before the
starting Saturday of week 1, we take 2 away from the ISO week 1 date (2 days
before Saturday is Thursday).
}
\details{
Format of ISO week is "YYYY-W__-num", where:
\itemize{
\item W__ is a range of W01 to W52 or W53 (depending on the year, there may be 53
defined weeks in a year, if there is a leap year).
\item num refers to the day of the ISO week, range 1 to 7, where 1 is Saturday.
e.g. 2023-W01-1 means the first day of the ISO week 1 in 2023.
Scotland uses different week start day, so the processing is different.
}
}
