% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{calc_uk_deaths_dashboard}
\alias{calc_uk_deaths_dashboard}
\title{Creates table of UK deaths for evergreen dashboard}
\usage{
calc_uk_deaths_dashboard(ew_deaths, scot_deaths, ni_deaths, cfg)
}
\arguments{
\item{ew_deaths}{Processed EW weekly deaths data frame, containing columns
region, cause, dor_year, dor_week, and week_ending.}

\item{cfg}{Config containing \code{year}, \code{week}, \code{scope} and \code{chart_years}.}

\item{scot_df}{Processed Scotland dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}
}
\value{
Data frame with UK death column.
}
\description{
Uses existing function "create_uk_dths_ctry_age_sex" to create
UK deaths data frame and modifies to evergreen dashboard specifications.
Filters, groups and pivots to create columns for UK deaths.
}
