% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_week_number_week_ending}
\alias{calc_week_number_week_ending}
\title{Calculate week number and week end date}
\usage{
calc_week_number_week_ending(data, col, current_year)
}
\arguments{
\item{data}{Deaths dataframe with year start date and year calculated (output
of calc_years_start_dates).}

\item{col}{String of date column name to use to create new columns from:
either "dor" or "dod".}

\item{current_year}{analytical year}
}
\value{
Deaths dataframe with extra columns for week number and week ending.
\code{year_start_date} column is dropped.
}
\description{
This function calculates the week number and week end date for a
given set of dates. The week number is based on the difference between each
date and a corresponding start date. It is calculated as the ceiling of the
number of weeks since the start date. Week end date is 6 days after the start
of that week (so difference in weeks minus 1 day).
}
\details{
When there is an existing column of \code{dor_week} (as in EW data), a
check will be done to compare the existing value of \code{dor_week} and the
created \code{week}. A warning will be given if there is a mismatch.
}
