% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_weeks_in_year}
\alias{calc_weeks_in_year}
\title{Calculate numbers of weeks in a year}
\usage{
calc_weeks_in_year(year)
}
\arguments{
\item{year}{year as numeric, e.g. 2024}
}
\value{
number of weeks in the year (52 or 53), as numeric
}
\description{
Calculate numbers of weeks in a year
}
