% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_year_end_date}
\alias{calc_year_end_date}
\title{Calculates end date for a year}
\usage{
calc_year_end_date(year, is_scot = FALSE)
}
\arguments{
\item{year}{year as numeric, e.g. 2024}

\item{is_scot}{is_scot logical, default is FALSE; for Scotland it is
standard ISO week}
}
\value{
date for end of the year as numeric, in format YYYYMMDD
}
\description{
Calculates end date for a year
}
