% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{calc_years_start_dates}
\alias{calc_years_start_dates}
\title{Calculate year start dates and their corresponding years for deaths data}
\usage{
calc_years_start_dates(data, col, is_scot = FALSE)
}
\arguments{
\item{data}{Deaths dataframe.}

\item{col}{String of date column name to use to create new columns from:
either "dor" or "dod".}

\item{is_scot}{logical, default is FALSE}
}
\value{
Deaths dataframe with extra column for year_start_date and year of
chosen column.
}
\description{
This function calculates the start date of the year for each
date in a given vector, and associates these dates with their corresponding
years. Gives a warning if \code{col} is "dor" and there are any 'missing' years
between max and min where there are no deaths.
}
\details{
Use a range from one year before to one year after to catch edge
cases of missing calendar years where there are no registrations and  weeks
at the start of the first calendar year that are in the previous registration
year, or at end of the last calendar year that are in the next
registration year.
}
