% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{calculate_asmr}
\alias{calculate_asmr}
\title{Calculate the age-standardised or age-specific mortality rate using the DHSC
methodology}
\usage{
calculate_asmr(deaths_age_pop_esp, grouping_vars)
}
\arguments{
\item{deaths_age_pop_esp}{Data frame with columns for \code{deaths}, \code{age_group},
\code{population}, \code{esp} and additional columns to be the \code{grouping_vars}. This df
needs to have all combinations of grouping variables i.e. 0 counts where
there are no deaths in a grouping.}

\item{grouping_vars}{Vector of strings}
}
\value{
Data frame
}
\description{
The input is a dataframe that must contain the columns:
\code{deaths}, \code{population}, \code{esp}, \code{agegroup_2} as well as the grouping columns:
\code{year}, \code{week}, \code{country}, \code{sex}.
The "All ages" data is filtered out (not used to calculate the ASMR, only
the age groups).
When the total deaths is < 10 ASMRs are not reliable and will therefore be
suppressed in the output.
Further info on the method can be found here:
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/methodologies/userguidetomortalitystatisticsjuly2017#death-rates-ratios-and-standardisation}{User guide to mortality statistics}.
\href{https://fingertips.phe.org.uk/static-reports/public-health-technical-guidance/Standardisation/DSRs.html}{DHSC directly standardised rates}
\href{https://github.com/ukhsa-collaboration/PHEindicatormethods/blob/master/R/phe_DSR.R}{PHE directly standardised rates (depricated)}

95\% Confidence intervals are calculated with the Byar's method with Dobson
method adjustment.
The rate is given per 100,000.
}
