% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{change_max_agegrp}
\alias{change_max_agegrp}
\title{Changes the maximum aggregations of the 5-year age bands to 100 and over}
\usage{
change_max_agegrp(data)
}
\arguments{
\item{data}{Dataframe with processed \code{agegroup_2} column.}
}
\value{
Dataframe with the changed \code{agegroup_2}.
}
\description{
Changes the maximum aggregations of the 5-year age bands to 100 and over
}
