% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{check_area_outputs_match}
\alias{check_area_outputs_match}
\title{Compare two tables to verify they contain the same results}
\usage{
check_area_outputs_match(
  tbl_baseline,
  tbl_to_check,
  areas,
  area_column = "Area of usual residence",
  death_column = "Number of deaths"
)
}
\arguments{
\item{tbl_baseline}{dataframe containing England and Wales death data for
comparison}

\item{tbl_to_check}{dataframe containing England and Wales death data to be
checked}

\item{areas}{vector of strings for each area to be specifically checked}

\item{area_column}{string, area column name in tables}

\item{death_column}{string, amount of deaths column name in tables}
}
\value{
dataframe containing the results from checking the table outputs
}
\description{
Compare two tables to verify they contain the same results
}
