% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{check_columns_present}
\alias{check_columns_present}
\title{Error if required columns not present}
\usage{
check_columns_present(data, required_cols)
}
\arguments{
\item{data}{Dataframe to check.}

\item{required_cols}{Vector of strings of the required columns.}
}
\value{
Error message stating which of the required columns are missing
}
\description{
Stops the execution of the function and returns an error
message stating which of the required columns are missing.
}
