% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{check_region_country}
\alias{check_region_country}
\title{Data validation: country and region}
\usage{
check_region_country(data)
}
\arguments{
\item{data}{Unprocessed deaths data.}
}
\value{
Error message if there are any mismatches.
}
\description{
Check contents of country and region and makes sure they
correctly correspond.
}
\details{
Checks regions correspond to England and Wales corresponds to Wales.
Non-residents correspond to NA in country. Note: as this data is unprocessed,
the region is "East of England", rather than "East". "East of England" should
be used in the publication.
}
