% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{compare_totals}
\alias{compare_totals}
\title{compare_totals}
\usage{
compare_totals(data, data_type, var, groupings, grouping_vars, overall)
}
\arguments{
\item{data}{dataframe containing number of deaths}

\item{data_type}{string identifying the type of data ("registrations",
"occurrences")}

\item{var}{string of the variable of interest}

\item{groupings}{vector groupings of var, e.g. Female, Male for sex.}

\item{grouping_vars}{vector of groupings to check sums for (age, sex, region)}

\item{overall}{string of the overall category e.g. All people for sex}
}
\value{
TRUE or FALSE if table exists and subtotals sum correctly.
}
\description{
Checks whether subgroups sum up to totals for all relevant tables in ONS
weekly deaths release. Intended for use in the production of a quality
report.
}
\details{
If the sum of subgroups on a table do not equal the total then the final
variable is set to FALSE. If the sum of subgroups on a table equal the total
then the final variable is set to YES.
}
