% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{convert_to_yearly_asmr}
\alias{convert_to_yearly_asmr}
\title{Convert weekly age-standardised mortality rate (ASMR) to yearly rate}
\usage{
convert_to_yearly_asmr(weekly_asmr, year)
}
\arguments{
\item{weekly_asmr}{Numeric. The weekly age-standardised mortality rate.}

\item{year}{Numeric. The year for which the conversion is being done. This is
used to determine if the year is a leap year.}
}
\value{
Numeric. The yearly age-standardised mortality rate.
}
\description{
This function takes the age-standardised mortality rate (ASMR)
calculated on a weekly basis and converts it to a yearly ASMR. This
annualisation of rates allows easier comparison of rate values between
different periods of time. More information can be found in the
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/methodologies/userguidetomortalitystatisticsjuly2017#death-rates-ratios-and-standardisation}{User guide to mortality statistics}.
}
\examples{
convert_to_yearly_asmr(weekly_asmr = 5, year = 2025)
convert_to_yearly_asmr(weekly_asmr = 5, year = 2024) # Leap year
}
