% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cause_geography.R
\name{count_dths_cause_geo}
\alias{count_dths_cause_geo}
\title{Create table containing death totals for each cause, per region, per week}
\usage{
count_dths_cause_geo(data)
}
\arguments{
\item{data}{Dataframe of the deaths dataset}
}
\value{
Dataframe containing death nums for causes, by region, per week
}
\description{
From the weekly deaths dataset, derives the number of deaths per
cause, on a per region level and all of this for each week in the year.
}
\details{
Developed using just England, Wales and Non-resident data, so
untested with Scotland and NI in mind, should this become relevant. It will
complete any missing cases with 0 values.
}
