% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cause_geography.R
\name{count_dths_cause_region}
\alias{count_dths_cause_region}
\title{Count deaths for each cause flag}
\usage{
count_dths_cause_region(data)
}
\arguments{
\item{data}{Dataframe containing the death data}
}
\value{
Dataframe containing all causes and their attributed number of deaths
per week, by region
}
\description{
For each cause flag, go through the process of grouping them
together by week and region so that we have a total number of deaths.
Then, join them all together and structure them so that we can have this
stored in a common cause and number column, rather than the original
structure of a column per flag.
}
