% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{count_dths_geo_age_sex}
\alias{count_dths_geo_age_sex}
\title{Count weekly deaths by area, age and sex}
\usage{
count_dths_geo_age_sex(df, current_year, current_week)
}
\arguments{
\item{df}{Data frame of death records with following columns:
dor, dod, country, region, agegrp_excess, sex and}

\item{current_year}{Year for analysis as numeric e.g. 2024}

\item{current_week}{Week for analysis as numeric or string e.g. 1}
}
\value{
A dataframe containing the summarised death counts.
}
\description{
Counts deaths by week, country, age group, sex
}
\details{
Filtering occurs after this, to only include the publication year and weeks
required up to and including the publication week. Filtering for dor took
place in the processing but an additional filtering needed to remove cases
where dod was in the previous year, which can only take place after year,
week and week_ending calculated based on dod.
0 counts are added for any missing combinations of factors. However, when 0
counts are added, week ending date is not completed so must be completed by
referring to the week number.
}
