% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{count_dths_geo_age_sex_pod}
\alias{count_dths_geo_age_sex_pod}
\title{Count weekly deaths by area, age, sex, and place of death}
\usage{
count_dths_geo_age_sex_pod(df, current_year, current_week)
}
\arguments{
\item{df}{Data frame of death records with following columns:
dor, dod, country, region, agegroup_5yr, sex and place_of_death.}

\item{current_year}{Year for analysis as numeric e.g. 2024.}

\item{current_week}{Week for analysis as numeric or string e.g. 1.}
}
\value{
A data frame containing the summarised death counts.
}
\description{
Counts deaths by week, country, region, age group, sex, and
place of death.
}
\details{
Filtering occurs to only include the publication year and weeks
required up to and including the publication week. Filtering for dor took
place in the processing but an additional filtering needed to remove cases
where dod was in the previous year, which can only take place after year,
week and week_ending calculated based on dod.
0 counts are added for any missing combinations of factors. However, when 0
counts are added, week ending date is not completed so must be completed by
referring to the week number.
}
