% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dths_imd.R
\name{count_imd}
\alias{count_imd}
\title{Count deaths by IMD quantile}
\usage{
count_imd(data, current_week, current_year, imd_lookup, quantile, country_var)
}
\arguments{
\item{data}{Deaths data frame.}

\item{current_week}{Week for analysis as numeric or string e.g. 1.}

\item{current_year}{Year for analysis as numeric e.g. 2024.}

\item{imd_lookup}{IMD lookup data frame.}

\item{quantile}{String of name of quantile for IMD, either "quintile" or
"decile".}

\item{country}{String of name of country, either "England" or "Wales".}
}
\value{
Formatted data frame.
}
\description{
Count deaths by IMD quantile
}
