% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{count_uk_dths_ctry_age_sex}
\alias{count_uk_dths_ctry_age_sex}
\title{Calculate UK counts of deaths by country, age groups and sex}
\usage{
count_uk_dths_ctry_age_sex(data, age_group = "agegrp_wide")
}
\arguments{
\item{data}{UK processed deaths dataframe.}

\item{age_group}{string of age group, default is "agegrp_wide"}
}
\value{
Dataframe with counts of deaths by week, country, agegroup and sex.
}
\description{
Takes the processed dataframe containing all deaths in the UK to
counts deaths by year, week, country (including UK and England/Wales/
Nonresidents), age group and sex. Totals are calculated for each variable
(except year and week).
}
