% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{create_area_cert_pod_delay}
\alias{create_area_cert_pod_delay}
\title{Registration delays by week of registration, selected death certification
type, selected area and place of death.}
\usage{
create_area_cert_pod_delay(data, area_col, cert_col)
}
\arguments{
\item{data}{Death registration data.}

\item{area_col}{Quoted name of column that refers to the area (either
"region" or "country").}

\item{cert_col}{Quoted name of column that refers to the certification type
(either "cert_type" or "cert_type_agg").}
}
\value{
A dataframe of delay data for all weeks by certification type, area,
place of death.
}
\description{
Uses England and Wales death registration data to create
number of deaths registered within 7 days, percentage of deaths registered
within 7 days, median delay and IQR of delay by certification type.
}
\details{
Counts certified by coroner are double counted between the
aggregation group and the with/without inquest so the total counts do not
equal 100\%.
The cert_col uses "cert_type_agg" to calculate the "Certified by coroner"
category
}
