% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_asmr_all_ppl_out_table}
\alias{create_asmr_all_ppl_out_table}
\title{Creates table where 'All people' asmr is unusually high or low}
\usage{
create_asmr_all_ppl_out_table(asmr_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing asmr data}
}
\value{
A dataframe with kableExtra styling applied
}
\description{
Creates the df where the 'All people' ASMR is outside the male
and female ASMRs. kableextra styling is applied ready to put in an Rmd.
}
\details{
Highlighted yellow as it is possible for this to happen without
being an error.
}
