% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_asmr_check_table}
\alias{create_asmr_check_table}
\title{Creates table of quality checks on ASMRs with kableextra styling}
\usage{
create_asmr_check_table(asmr_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing asmr data}
}
\value{
A dataframe with a three rows with an Outcome column of either TRUE, FALSE,
Check or 'Check if table exists' with styling applied
}
\description{
Creates the dataframe with kable styling applied to check if ASMR is between
limits, UK value is between min and max country values and All people value
is between male and female
}
