% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_asmr_out_lims_table}
\alias{create_asmr_out_lims_table}
\title{Create table where asmr is ouside upper and lower bound values}
\usage{
create_asmr_out_lims_table(asmr_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing asmr data}
}
\value{
A dataframe with kableExtra styling applied
}
\description{
Creates the df where the ASMR is outside the upper and lower
limits. kableextra styling is applied ready to put in an Rmd.
}
