% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_dashboard_data}
\alias{create_dashboard_data}
\title{Creates table for evergreen dashboard}
\usage{
create_dashboard_data(
  ew_deaths,
  exp_deaths_sql,
  scot_deaths,
  ni_deaths,
  cfg,
  num_yr_to_keep
)
}
\arguments{
\item{ew_deaths}{Processed EW weekly deaths data frame, containing columns
region, cause, dor_year, dor_week, and week_ending.}

\item{exp_deaths_sql}{Data frame containing expected death data including
area, age, sex.}

\item{cfg}{Config containing \code{year}, \code{week}, \code{scope} and \code{chart_years}.}

\item{num_yr_to_keep}{Numeric for how many years of data to keep.}

\item{scot_df}{Processed Scotland dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}
}
\value{
Data frame with columns for England and Wales deaths (by country
and cause), expected_deaths and UK deaths.
}
\description{
Creates table for evergreen dashboard, calling 3 sub-functions
to bring in England and Wales deaths (by country and cause), expected_deaths
and UK deaths. If there are more than 2 calendar years in the data,
rows are dropped to result in only a specific number of years of data.
}
