% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{create_delay_area_pod_cert_sum}
\alias{create_delay_area_pod_cert_sum}
\title{Registration delays by week of registration, death certification type, area
and place of death.}
\usage{
create_delay_area_pod_cert_sum(data)
}
\arguments{
\item{data}{Death registration data.}
}
\value{
Data frame of combinations for delays by week, certification type,
area and place of death.
}
\description{
Registration delays by week of registration, death certification type, area
and place of death.
}
