% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{create_delay_summary}
\alias{create_delay_summary}
\title{Registration delays by groups of interest (England and Wales)}
\usage{
create_delay_summary(data, groupings)
}
\arguments{
\item{data}{Death registration data.}

\item{groupings}{Vector of strings for column names to group by.}
}
\value{
A dataframe of delay data.
}
\description{
Uses England and Wales death registration data to create
number of deaths registered within 7 days, percentage of deaths registered
within 7 days, median delay and IQR of delay.
}
