% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{create_delay_summary_certtype}
\alias{create_delay_summary_certtype}
\title{Registration delays by week of registration and death certification type
(England and Wales)}
\usage{
create_delay_summary_certtype(data)
}
\arguments{
\item{data}{Death registration data.}
}
\value{
A dataframe of delay data for all weeks by certification type.
}
\description{
Uses England and Wales death registration data to create
number of deaths registered within 7 days, percentage of deaths registered
within 7 days, median delay and IQR of delay by certification type.
}
\details{
Counts certified by coroner are double counted between the
aggregation group and the with/without inquest so the total counts do not
equal 100\%.
}
