% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_due_involve_rows_table}
\alias{create_due_involve_rows_table}
\title{create_due_involve_rows_table}
\usage{
create_due_involve_rows_table(cause_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing the
relevant data.}
}
\value{
A dataframe containing the rows from the causes table where deaths due to a
cause exceed the deaths involving the cause.
}
\description{
Creates the dataframe with kable styling applied displaying the rows where
the due to number of deaths is greater than the involving number of deaths.
If there are no rows, nothing is returned.
}
