% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_due_involve_table}
\alias{create_due_involve_table}
\title{create_due_involve_table}
\usage{
create_due_involve_table(cause_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing the
relevant data.}
}
\value{
A dataframe with a single row with an Outcome column of either TRUE, FALSE or
'Check if table exists'
}
\description{
Creates the dataframe with kable styling applied to show whether the number
of rows due to a cause is less than or equal to the number of rows involving
a cause
}
