% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_occ_current_previous_years.R
\name{create_ew_occurrences_table}
\alias{create_ew_occurrences_table}
\title{Create final occurrences table by combining current and previous occurrences
data}
\usage{
create_ew_occurrences_table(df1, df2, imd, config)
}
\arguments{
\item{df1}{data frame of deaths in England and Wales filtered to current year}

\item{df2}{data frame of deaths in England and Wales not filtered}

\item{imd}{list containing items england and wales, the imd lookups}

\item{config}{list which includes items year (current year - numeric e.g.
2024), week (analysis week - numeric e.g. 24), previous_year (previous_year,
numeric e.g. 2023)}
}
\value{
A data-frame combining two years of summarised death counts.
}
\description{
Create final occurrences table by combining current and previous occurrences
data
}
