% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{create_excess_age_groups}
\alias{create_excess_age_groups}
\title{Creates excess death age groups}
\usage{
create_excess_age_groups(data, age_column)
}
\arguments{
\item{data}{dataframe containing data to be reorganised}

\item{age_column}{string of the age column's name}
}
\value{
dataframe with new column agegrp_excess
}
\description{
Creates excess death age groups
}
\details{
Converts esp/5 year bands to excess death bands, these are 0 to 29,
30 to 44, 45-49, 50-59 etc.
}
