% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_levels.R
\name{create_factor_levels}
\alias{create_factor_levels}
\title{Creates a list object of all the factor levels in correct order per column}
\usage{
create_factor_levels()
}
\value{
List object of string vectors of factor levels
}
\description{
Creates a list object of all the factor levels in correct order per column
}
