% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_lengths_table}
\alias{create_lengths_table}
\title{Create df with info about whether table lengths are as expected}
\usage{
create_lengths_table(
  reg_table,
  occ_table,
  cause_table,
  delay_table,
  excess_table,
  uk_table,
  asmr_table,
  weeks,
  scope,
  previous_year
)
}
\arguments{
\item{reg_table}{data frame containing death registration data}

\item{occ_table}{data frame containing death occurrences data}

\item{cause_table}{data frame containing cause of death data}

\item{delay_table}{data frame containing death registration delays data}

\item{excess_table}{data frame containing excess death data}

\item{uk_table}{data frame containing UK death registration data}

\item{asmr_table}{data frame containing UK ASMR data}

\item{weeks}{week for analysis as numeric or string e.g. 1}
}
\value{
A data frame containing the table number, the table's data type,
the number of rows in table, the number of rows that are expected in the
table and a matching column to indicate if the table is as expected.
}
\description{
Create df with info about whether table lengths are as expected
}
