% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_log.R
\name{create_run_log}
\alias{create_run_log}
\title{Create a run log}
\usage{
create_run_log(cfg, timestamp)
}
\arguments{
\item{cfg}{a list which includes,  week_no, publication_date,
years_for_averages_finalised, years_for_averages_provisional and details of
server and tables used for death data}

\item{time_stamp}{Time as character string, unified with other saved outputs.}
}
\description{
Produces a .txt file that is saved to the outputs folder and
contains information about the time the RAP is run, user running the code,
location of the code, week and years used, and server and tables used.
}
