% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{create_subtotals_table}
\alias{create_subtotals_table}
\title{create_subtotals_table}
\usage{
create_subtotals_table(
  ew_registrations,
  ew_occurrences,
  ew_registrations_by_cause,
  ew_reg_delays_cert_pod_area,
  uk_excess_deaths,
  uk_registrations,
  week,
  scope
)
}
\arguments{
\item{ew_registrations}{dataframe of death registration data}

\item{ew_occurrences}{dataframe of death occurrence data}

\item{ew_registrations_by_cause}{dataframe of death registration data by
cause of death}

\item{uk_excess_deaths}{dataframe of death registration data by excess deaths}

\item{uk_registrations}{dataframe containing death registration data for
England, Wales, Scotland, Northern Ireland and the uK.}
}
\value{
A knitr::kable for a rmd file
}
\description{
Function for creating table checking subtotals sum to relevant
total for the relevant tables in weekly deaths
}
