% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{create_uk_df}
\alias{create_uk_df}
\title{Process UK country ages and filter to period of interest}
\usage{
create_uk_df(ew_df, scot_df, ni_df, current_year)
}
\arguments{
\item{ew_df}{Processed England and Wales dataframe.}

\item{scot_df}{Processed Scotland dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}

\item{current_year}{Analysis year as numeric in format YYYY.}
}
\value{
UK-level dataframe or error if there is more than one year of data.
}
\description{
Processes the England/Wales (EW), Scotland and Northern Ireland
dataframes to be in publishing ready age groups (e.g. "15 to 19").
Adds a column of 1 for \code{deaths} for England and Wales as the data is
record-level. Drops Scotland counts for deaths involving covid to leave "all
deaths" as covid are duplicated counts. Uses Scotland week definition if
necessary.
Binds all deaths records together.
}
\details{
The import and selection of single year of England + Wales data
already should already be filtered to the single year during the processing.
}
