% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{create_uk_dths_ctry_age_sex}
\alias{create_uk_dths_ctry_age_sex}
\title{Combines England and Wales, Scotland and Northern Ireland and calculate UK
counts of deaths by country, age groups and sex}
\usage{
create_uk_dths_ctry_age_sex(
  ew_df,
  scot_df,
  ni_df,
  current_year,
  current_week,
  scope,
  age_group = "agegrp_wide"
)
}
\arguments{
\item{ew_df}{Processed England and Wales dataframe.}

\item{scot_df}{Processed Scotland dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}

\item{current_year}{Analysis year as numeric in format YYYY.}

\item{current_week}{Week for analysis as numeric or string e.g. 1}

\item{scope}{string of "UK" or "EW",}

\item{age_group}{string of age group, default is "agegrp_wide"}
}
\value{
Dataframe of UK counts of deaths by country, age groups and sex.
}
\description{
Filters out any data outside of the analysis time period and
combines datasets for England and Wales, Scotland and Northern Ireland. Then
counts deaths by week, country, age and sex. If no deaths occur in a group,
counts of 0 are added. If the scope is EW then there are no rows for UK, NI,
Scot for the latest week.
}
\details{
Adds a column of 1 for \code{deaths} for England and Wales as the data is
record-level. Drops Scotland counts for deaths involving covid to leave "all
deaths" as covid are duplicated counts.
}
