% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{derive_max_reg_date}
\alias{derive_max_reg_date}
\title{Determine what the maximum date should be for considered registration dates.}
\usage{
derive_max_reg_date(cfg, is_scot = FALSE)
}
\arguments{
\item{cfg}{Config as named list containing year and week named elements.
year as integer in format YYYY and week as string.}

\item{is_scot}{logical, default is FALSE}
}
\value{
character containing the max date in the format YYYYMMDD
}
\description{
Determine what the maximum date should be for considered registration dates.
}
