% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{derive_years_week_numbers}
\alias{derive_years_week_numbers}
\title{Derive week and year information for dates of death occurrences or
registration}
\usage{
derive_years_week_numbers(data, current_year, col = "dor", is_scot = FALSE)
}
\arguments{
\item{data}{Dataframe with the columns: \code{dor} and \code{dod}, which
contain dates in YYYYMMDD format and are integer type.}

\item{current_year}{analytical year}

\item{col}{String of either "dor" or "dod", default is "dor".}

\item{is_scot}{logical, default is FALSE}
}
\value{
The input dataframe with 3 new columns: year, week (week number) and
week_ending, for either registration or occurrence date.
}
\description{
This function takes a dataframe and calculates the start date
of the year and the publication year for each date. It then calculates the
week number based on the derived year start date. This is done for either
\code{dor} (registration) or \code{dod} (occurrences).
Note: The week and year number derivation is based on a modified version of
the ISO 8601 standard. In this modification, the year's start date is
considered to be the Saturday preceding the first week of the year. The
first week of the year is defined as the week containing the first Thursday
of the year. Week end number is also added.
}
