% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{display_asmr_warning_text}
\alias{display_asmr_warning_text}
\title{display_asmr_warning_text}
\usage{
display_asmr_warning_text(asmr_table)
}
\arguments{
\item{data}{The R dataframe in the global environment containing asmr data}
}
\value{
char with knitr styling applied.
}
\description{
Creates the text to display when there are rows in the asmr
table which fail the checks.
}
