% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{due_involve_check}
\alias{due_involve_check}
\title{due_involve_check}
\usage{
due_involve_check(data)
}
\arguments{
\item{data}{The R dataframe in the global environment containing the
relevant data.}
}
\value{
A dataframe "due_involve_df" within the local function environment. Nothing
will be saved to the global environment by running the function alone.
}
\description{
Comparing the counts of deaths involving a condition with the counts of
deaths due to a condition for ONS weekly deaths release. Intended for use in
the production of a quality report.
}
\details{
This function compares the counts of deaths involving a condition with the
counts of deaths due to a condition. If the count of deaths involving a
condition is less than or equal to the count of deaths due to a condition
by week number and area of residence, then the row of interest is kept in
the final dataframe.
}
\examples{
nrow(due_involve_check(ew_registrations_by_cause))

}
