% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_dths.R
\name{excess_deaths_5ya}
\alias{excess_deaths_5ya}
\title{Compute excess deaths using five year average}
\usage{
excess_deaths_5ya(
  weekly_df,
  annual_df,
  years_weekly,
  years_annual,
  current_week,
  current_year
)
}
\arguments{
\item{weekly_df}{Provisional deaths dataset with weekly data.}

\item{annual_df}{Finalised annual deaths dataset containing years of
interest.}

\item{years_weekly}{Numeric vector, specifying which years contained in
weekly dataset should be used for computing five year average.}

\item{years_annual}{Numeric vector, specifying which years contained in
annual dataset should be used for computing five year average.}

\item{current_week}{Analytical week number}

\item{current_year}{Analytical year as full number, e.g. 2024}
}
\value{
data frame with total deaths for this year, five year average and
percentage change from five year average; by week, country and place of death
}
\description{
Compute excess deaths using five year average
}
