% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_processing.R
\name{filter_before_year_start}
\alias{filter_before_year_start}
\title{Drop records with registration date before the start of the analysis year}
\usage{
filter_before_year_start(data, current_year, is_scot = FALSE)
}
\arguments{
\item{data}{Deaths dataframe with \code{dor} column.}

\item{current_year}{analytical year as numeric e.g. 2024}

\item{is_scot}{logical, default is FALSE}
}
\value{
Deaths dataframe with records before the start of the analysis year.
}
\description{
Drop records with registration date before the start of the analysis year
}
