% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_agegrp_excess}
\alias{flag_agegrp_excess}
\title{Flag age group (for excess publication)}
\usage{
flag_agegrp_excess(data, age_col)
}
\arguments{
\item{data}{Data frame with column for age in years.}

\item{age_col}{Unquoted name of column with numeric age (in years).}
}
\value{
Data frame with added agegrp_wide column.
}
\description{
Flag age group (for excess publication)
}
