% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_cause}
\alias{flag_cause}
\title{Flag cause of death}
\usage{
flag_cause(data, new_col, codes, due_to = FALSE)
}
\arguments{
\item{data}{Deaths data frame.}

\item{new_col}{Unquoted name of new column to create.}

\item{codes}{Vector of strings of ICD10 codes.}

\item{due_to}{Boolean of whether the flag is for deaths due to cause codes
(TRUE) or involving cause codes (FALSE). Default is FALSE. Error if wrong
arg given.}
}
\value{
Data frame with extra column.
}
\description{
Uses vector of ICD10 codes to search columns for matches. If
due_to is set to FALSE, it will search columns beginning with "fic"
(including underlying cause, "fic10und"), and if due_to is set to TRUE,
it will only look at "fic10und". A new column is created with TRUE or FALSE
as to whether there are any matches.
}
\details{
For non-COVID-19 codes, 4-digit codes will match to their 3-digit
header: e.g., "J182" would flag in "J18". COVID-19 codes require a 4-digit
code match.
}
