% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_cert_type_agg}
\alias{flag_cert_type_agg}
\title{Creates aggregated certification type}
\usage{
flag_cert_type_agg(data, cert_column)
}
\arguments{
\item{data}{Data frame with following columns: \code{cert_type}.}

\item{cert_column}{Unquoted string of the \code{cert_type} column's name.}
}
\value{
Data frame with new column \code{cert_type_agg}.
}
\description{
Creates aggregated certification type
}
\details{
Converts \code{cert_type} to add an aggregated \code{cert_type} for
'certified by coroner'.
}
