% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_cert_type_groups}
\alias{flag_cert_type_groups}
\title{Derive certification type}
\usage{
flag_cert_type_groups(data, cert_column)
}
\arguments{
\item{data}{Annual deaths data frame with following columns: \code{certtype}.}

\item{cert_column}{Unquoted string of the \code{cert_type} column's name.}
}
\value{
Data frame with new \code{cert_type} column, specified as one of:
Certified by doctor, Certified by coroner with inquest,
Certified by coroner with no inquest, Uncertified/Unknown or
Certified by medical examiner.
}
\description{
Derive certification type
}
