% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_country_ew}
\alias{flag_country_ew}
\title{Flag country (England and Wales)}
\usage{
flag_country_ew(data)
}
\arguments{
\item{data}{Data frame with ctryr column.}
}
\value{
Data frame with added country column
}
\description{
Uses \code{ctryr} variable containing country classification codes
to flag entries with "England" (921), "Wales" (924) or "Nonresident"
(all others).
}
