% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_country_region}
\alias{flag_country_region}
\title{Convert postcode into country and region information}
\usage{
flag_country_region(data, nspl)
}
\arguments{
\item{data}{data frame with pcdr column specifying postcode of residence}

\item{nspl}{data frame with pcd column with postcodes and gor column with
codes for regions}
}
\value{
original data frame with additional region and country columns, but
without pcdr column
}
\description{
Convert postcode into country and region information
}
