% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_place_of_death_annual}
\alias{flag_place_of_death_annual}
\title{Convert place of deaths into names for annual files}
\usage{
flag_place_of_death_annual(data)
}
\arguments{
\item{data}{annual deaths data frame with following columns: cestrss,
esttyped, nhsind}
}
\value{
data frame with new place_of_death column, specified as one of: Home,
Hospital, Care home, Hospice or Elsewhere; Columns cestrss, esttyped, nhsind
have been removed
}
\description{
Convert place of deaths into names for annual files
}
