% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_place_of_death_other}
\alias{flag_place_of_death_other}
\title{Flag place of death for UK excess mortality}
\usage{
flag_place_of_death_other(data, data_type)
}
\arguments{
\item{data}{Data frame or England + Wales data.}

\item{data_type}{String of "finalised" or "provisional", describing what type
of data is supplied. An error will be given if the input does not match one
of these strings.}
}
\value{
Data frame with place_of_death_other column.
}
\description{
Creates place_of_death_other column with values: "Home",
"Hospital", "Care home", and "Other". If data is "provisional",
\code{place_of_death} column is used to derive  \code{place_of_death_other}. If data is
"finalised", \code{place_of_death_other} is derived from \code{cestrss}, \code{nhsind}, and
\code{esttyped} as there is no \code{place_of_death} column.
}
\details{
\code{cestrss} has values of "H" for home, otherwise numeric to denote
communal establishment. \code{place_of_death} is dropped for the "provisional"
processing.
}
