% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{flag_weekly_causes}
\alias{flag_weekly_causes}
\title{Flag cause of death for weekly deaths publication.}
\usage{
flag_weekly_causes(data)
}
\arguments{
\item{data}{Weekly deaths data frame.}
}
\value{
Data frame with extra columns.
}
\description{
Uses ICD10 codes to create new columns with boolean entries
(TRUE or FALSE) for whether deaths are due to or involving causes (diseases
of the respiratory system, influenza or pneumonia, and COVID-19). Due to is
is TRUE if the code is present in the "fic10und" column, involving is TRUE
if the code is in any of the fic10 columns.
}
\details{
Respiratory system codes are J00 to J99 (inc. 4-digit codes).
Influenza or pneumonia codes are J09 to J18 (inc. 4-digit codes).
COVID-19 codes are U071, U072, U099 and U109.
}
