% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{format_asmr}
\alias{format_asmr}
\title{Format the age-standardised mortality rate (ASMR) for publication}
\usage{
format_asmr(data)
}
\arguments{
\item{data}{ASMR dataframe (output of \code{phe_age_rates}).}
}
\value{
Dataframe ready for publication or error if there would be any NA in
the output.
}
\description{
Makes the country and sex variables factors to sort them into
the correct order. Then renames all required columns and drops other columns.
Due to unreliability, if the deaths are < 10, the ASMR is suppressed and the
ASMR column will be a character column.
}
