% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{format_dashboard}
\alias{format_dashboard}
\title{Format data ready for Evergreen dashboard}
\usage{
format_dashboard(data, cfg, round_to = 0)
}
\arguments{
\item{data}{Data frame containing deaths registered, expected deaths and
deaths by cause (both due and involving COVID-19 and influenza or pneumonia)
for England and Wales.}

\item{cfg}{Config containing \code{publication_date}.}

\item{round_to}{Integer used to round expected death to this decimal place.}
}
\value{
Formatted data frame containing deaths registered, expected deaths
and deaths by cause. Order columns and rows to dashboard specification.
}
\description{
Creates a publication date column (date type) using the config
variable in the format "YYYY-MM-DD". Any column starting with
"Expected deaths" is rounded to specified number of decimal places.
The \code{week_ending} column is also converted into format "YYYY-MM-DD".
The data frame is sorted by the \code{week_ending} and only required columns
are selected and named as appropriate.
}
\details{
The format of the output dataframe should be agreed with publishing
as it is used in their pipeline to produce the Evergreen dashboard. No
changes without communication.
}
