% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{format_delay_summary}
\alias{format_delay_summary}
\title{Format registration delay table}
\usage{
format_delay_summary(df)
}
\arguments{
\item{df}{Unformatted death registration delay df.}
}
\value{
Formatted dataframe of death registration delay data.
}
\description{
Adds week end date column after week number,
sorts by week number descending and renames columns.
}
