% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_delays.R
\name{format_delay_summary_certtype}
\alias{format_delay_summary_certtype}
\title{Format registration delay table with certification type included}
\usage{
format_delay_summary_certtype(df)
}
\arguments{
\item{df}{Unformatted death registration delay dataframe.}
}
\value{
Formatted dataframe of death registration delay data by certification
type.
}
\description{
Add zero counts, renames, relocates and orders columns.
}
