% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cause_geography.R
\name{format_dths_cause_geo}
\alias{format_dths_cause_geo}
\title{Handle final formating for deaths per cause, per region, by week table}
\usage{
format_dths_cause_geo(data)
}
\arguments{
\item{data}{Dataset containing the output from the
create_tbl_dths_per_cause_and_regn_by_wk() function}
}
\value{
Dataset containing the formatted output from the
create_tbl_dths_per_cause_and_regn_by_wk()
}
\description{
Perform final formatting on the table to get the data into a
more publishable state.
Such as:
• Factorising relevant columns
• Ordering of data
• Ordering of columns
• Renaming column to something more human-readable
}
