% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{format_dths_geog_age_sex_pod_imd}
\alias{format_dths_geog_age_sex_pod_imd}
\title{Format weekly deaths by country/region, age, sex, and place of death.}
\usage{
format_dths_geog_age_sex_pod_imd(data)
}
\arguments{
\item{data}{Data frame with week, country_region, agegrp_5yr, sex, imd,
place_of_death, and count columns.}
}
\value{
Formatted data frame.
}
\description{
Format weekly deaths by country/region, age, sex, and place of death.
}
