% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_dths.R
\name{format_excess_deaths_5ya}
\alias{format_excess_deaths_5ya}
\title{Format five year average table}
\usage{
format_excess_deaths_5ya(df)
}
\arguments{
\item{df}{output of excess_deaths_5ya function}
}
\value{
Formatted data frame.
}
\description{
Format five year average table
}
