% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{format_excess_deaths_model}
\alias{format_excess_deaths_model}
\title{Format weekly deaths by country, age, sex, and place of death.}
\usage{
format_excess_deaths_model(data)
}
\arguments{
\item{data}{Data frame with week, country_region, agegrp_excess, sex,
place_of_death, and count columns.}
}
\value{
Formatted data frame.
}
\description{
Format weekly deaths by country, age, sex, and place of death.
}
