% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{format_qa_table}
\alias{format_qa_table}
\title{Function to add background colour to matching column}
\usage{
format_qa_table(df)
}
\arguments{
\item{df}{QA data frame - containing columns Area level checked and match}
}
\value{
data frame with coloured background to indicate if QA is matching,
not matching or needs checking
}
\description{
Function to add background colour to matching column
}
