% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{format_uk_dths_ctry_age_sex}
\alias{format_uk_dths_ctry_age_sex}
\title{Format UK counts of deaths by country, agegroups and sex}
\usage{
format_uk_dths_ctry_age_sex(data)
}
\arguments{
\item{data}{Output dataframe of \code{create_uk_dths_ctry_age_sex}.}
}
\value{
Dataframe of UK counts of deaths by country, agegroups and sex,
formatted for publishing.
}
\description{
Format UK counts of deaths by country, agegroups and sex
}
