% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{generate_report}
\alias{generate_report}
\title{Generate a validation report for a particular column}
\usage{
generate_report(col, valid_func)
}
\arguments{
\item{col}{Column to apply validation to, in format data$column.}

\item{valid_func}{The validation function to apply to the column.}
}
\value{
List output of validation report.
}
\description{
Generate a validation report for a particular column
}
