% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{get_delays_subgroup_info}
\alias{get_delays_subgroup_info}
\title{Get subgroup information for delays}
\usage{
get_delays_subgroup_info(data, x, topic)
}
\arguments{
\item{data}{Data frame of delays.}

\item{x}{String of data frame name.}

\item{topic}{String of grouping of interest, e.g., region, place, cert_type.}
}
\value{
A list with named elements var, groupings, vars, overall, data
}
\description{
Uses England and Wales delays table to check subtotals against
overall total.
}
