% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{get_excess_subgroup_info}
\alias{get_excess_subgroup_info}
\title{get_excess_subgroup_info}
\usage{
get_excess_subgroup_info(data, x, topic)
}
\arguments{
\item{data}{dataframe of death data}

\item{x}{string of dataframe name}

\item{topic}{string grouping of interest e.g. region, age, sex, place}
}
\value{
a list with named elements var, groupings, vars, overall, data
}
\description{
Function forew registrations by cause table for creating var the
variable that subtotals will be checked for, groupings that make up the var,
vars that will be used for the breakdowns, overall for what the groupings
should sum to and data which is data filtered as required.
}
