% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{get_rows_asmr_all_ppl_out_m_f}
\alias{get_rows_asmr_all_ppl_out_m_f}
\title{Gets rows where 'All people' ASMR is outside male and female ASMRs}
\usage{
get_rows_asmr_all_ppl_out_m_f(asmr_df)
}
\arguments{
\item{asmr_df}{dataframe with cols \verb{ASMR per 100,000}, \code{Country}, \code{Sex},
\verb{Week number}}
}
\value{
dataframe
}
\description{
Creates a df of rows where the all people asmr table fails the
check of being between the male and female ASMRS
}
\details{
It is possible for this to happen without error due to the weighting
of populations by sex and ESP.
}
