% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{get_rows_asmr_outside_limits}
\alias{get_rows_asmr_outside_limits}
\title{Gets rows where asmr value is not between lower and upper limits}
\usage{
get_rows_asmr_outside_limits(asmr_df)
}
\arguments{
\item{asmr_df}{dataframe with cols \verb{ASMR per 100,000},
\verb{Lower confidence limit} and \verb{Upper confidence limit}}
}
\value{
dataframe
}
\description{
Creates a df of rows where the asmr table fails the check of
asmr being between lower and upper limits
}
