% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{get_rows_uk_asmr_out_ctrys}
\alias{get_rows_uk_asmr_out_ctrys}
\title{Gets rows where UK ASMR is outside individual countries ASMRs}
\usage{
get_rows_uk_asmr_out_ctrys(asmr_df)
}
\arguments{
\item{asmr_df}{dataframe with cols \verb{ASMR per 100,000}, \code{Country}, \code{Sex},
\verb{Week number}}
}
\value{
dataframe
}
\description{
Creates a df of rows where the asmr table fails the check of
the UK asmr being between min and max of individual countries of the UK
}
\details{
It is possible for this to happen without error due to the weighting
of populations by sex and ESP.
}
