% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_by_summarise}
\alias{group_by_summarise}
\title{Group by and summarise}
\usage{
group_by_summarise(data, grouping_cols)
}
\arguments{
\item{data}{Data frame.}

\item{grouping_cols}{Vector of strings of column names to group by.}
}
\value{
Data frame with counts column (ungrouped).
}
\description{
Groups by specified columns, summarises (counts). All
combinations of factor variable levels will be returned, even with no
instances.
}
