% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{group_count}
\alias{group_count}
\title{Count deaths records by group}
\usage{
group_count(df, groupings = c())
}
\arguments{
\item{df}{Data frame with 1 row equivalent to 1 death. Has to include the
desired groupings, specified in groupings argument.}

\item{groupings}{Vector of column names to group by as strings.}
}
\value{
Data frame with number of deaths by desired groupings.
}
\description{
Counts deaths records by group, excludes non-residents if groupings include
country, excludes non-residents and Wales if groupings include regions. If
neither country or region groupings are specified, area is assumed to  be
England, Wales and non-residents.
}
