% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{group_count_all_geo}
\alias{group_count_all_geo}
\title{Count death records by group and area in England and Wales}
\usage{
group_count_all_geo(df, groupings = c())
}
\arguments{
\item{df}{Data frame with 1 row equivalent to 1 death. Has to include region
and country columns and the desired groupings, specified in groupings
argument.}

\item{groupings}{Vector of column names to group by as strings.}
}
\value{
Data frame with number of deaths by area and specified groupings.
}
\description{
Counts deaths records by specified groupings for following geographies:
England, Wales and non-residents, England, Wales, English regions.
}
