% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_esp}
\alias{import_esp}
\title{Imports European Standard Population (ESP) lookup}
\usage{
import_esp(cfg_list)
}
\arguments{
\item{cfg_list}{List containing following fields:
\itemize{
\item{"server"}{SQL server name as a string}
\item{"db_pops"}{SQL population database name as a string}
\item{"tbl_esp"}{ESP table name as a string}
\item{"vars_esp"}{ESP variables as a string}
}}
}
\value{
ESP data frame.
}
\description{
Imports European Standard Population (ESP) lookup
}
