% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_ew}
\alias{import_ew}
\title{Import data for England and Wales from the provisional files}
\usage{
import_ew(cfg)
}
\arguments{
\item{cfg}{list containing following fields:
\itemize{
\item{"year"}{analysis year as numeric e.g. 2024}
\item{"provisional_years"}{list of 4-digit years of provisional data}
\item{"var_ew"}{character vector with variable names for provisional data}
\item{"var_cause"}{character vector with variable names for columns with
cause codes}
\item{"server"}{SQL server name as a string}
\item{"db_weekly_dths"}{database for provisional deaths data}
\item{"tbl_ew"}{table name for provisional deaths data}
}}
}
\value{
data frame with deaths data
}
\description{
Import of provisional death data to be joined with finalised
death data.
}
\details{
in current form this function is built to use with registration
data. The start date is 1st Jan of the first provisional year, as this
will be joined with finalised data up to 31st Dec of the previous year.
The config is updated in the global scope to include ew_file_final,
which can be used for reporting.
}
