% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_ew_final}
\alias{import_ew_final}
\title{Import finalised data for England and Wales}
\usage{
import_ew_final(cfg)
}
\arguments{
\item{cfg}{list containing following fields:
\itemize{
\item{"finalised_years"}{numeric vector of years for finalised data (YYYY)}
\item{"var_cause"}{character vector with variable names for columns with
cause codes}
\item{"server"}{SQL server name as a string}
\item{"db_dths"}{database for annual (finalised) deaths data}
\item{"tbl_ew_2018on"}{table name for annual deaths data from 2018}
\item{"tbl_ew_2001_2017"}{table name for annual data from 2001 to 2017}
\item{"var_ew_final"}{character vector of variables from finalised death data
tables}
}}
}
\value{
data frame with deaths data
}
\description{
Imports finalised death data to be joined with provisional
weekly deaths data
}
\details{
ew_file_used_final_report is saved to cfg for reporting. The end
date is 31st of Dec for the last finalised year as this will be joined with
provisional data from 1st Jan of the following year.
The config is updated in the global scope to include ew_file_used_final,
which can be used for reporting.
}
