% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_excess_death_data_sql}
\alias{import_excess_death_data_sql}
\title{Imports excess deaths data from SQL}
\usage{
import_excess_death_data_sql(cfg_list)
}
\arguments{
\item{config}{list containing fields for deaths server, excess deaths
database and excess deaths table.}
}
\value{
dataframe containing excess death data including area, age, sex
}
\description{
Imports excess deaths data from SQL
}
\details{
Reads data from whole excess deaths data table from sql
}
