% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_imd}
\alias{import_imd}
\title{Import Index of Multiple Deprivation (IMD) data}
\usage{
import_imd(cfg_list, country)
}
\arguments{
\item{cfg_list}{\itemize{
\item{"server"}{SQL server name}
\item{"db_ref}{SQL reference database name}
\item{"tbl_imd_england}{table name for IMD lookup for England}
\item{"tbl_imd_wales}{table name for IMD lookup for Wales}
\item{"var_imd_england"}{string of variables to import for England}
\item{"var_imd_wales"}{string of variable to import for Wales}}}

\item{country}{String value of country; either "england" or "wales".}
}
\value{
Data frame for IMD lookup table or error if \code{country} is not
"england" or "wales".
}
\description{
Import Index of Multiple Deprivation (IMD) data
}
