% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_nspl}
\alias{import_nspl}
\title{Imports nspl data for England and Wales}
\usage{
import_nspl(cfg)
}
\arguments{
\item{cfg}{list containing following fields:
\itemize{
\item{"server"}{SQL server name as a string}
\item{"db_ref}{SQL reference data base name as a string}
\item{"tbl_nspl"}{nspl table name as a string}
}}
}
\value{
nspl data frame with postcodes (pcd), region codes (gor) and  lower
layer output areas (lsoa11) for England and Wales.
}
\description{
Imports nspl data for England and Wales
}
