% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_poisson}
\alias{import_poisson}
\title{Imports Poisson distribution lookup}
\usage{
import_poisson(cfg_list)
}
\arguments{
\item{cfg_list}{list containing following fields:
\itemize{
\item{"server"}{SQL server name as a string}
\item{"db_pops"}{SQL population database name as a string}
\item{"tbl_poisson"}{Poisson distribution table name as a string}
\item{"vars_poisson"}{Poisson distribution variables as a string}
}}
}
\value{
Poisson distribution as data frame.
}
\description{
The Poisson distribution is used to calculate confidence
intervals e.g. for age-standardised mortality rate (ASMR).
}
