% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_scot}
\alias{import_scot}
\title{Import Scotland death data}
\usage{
import_scot(cfg_list)
}
\arguments{
\item{cfg_list}{list containing following fields:
\itemize{
\item{"scope"}{analysis scope as string e.g. "EW"}
\item{"year"}{analysis year as numeric e.g. 2024}
\item{"week"}{analysis week as numeric e.g. 15}
\item{"server"}{SQL server name as a string}
\item{"db_scot_ni"}{database name as string}
\item{"scot_ni_tbl"}{current scot and ni data table name as string}
\item{tbl_scot}{Scot data table name as string}
\item{var_scot}{character vector with variable names for provisional data}
}}
}
\value{
Deaths data frame with columns defined in the config.
}
\description{
This function imports death data for Scotland. If cfg$scope is
"UK" the latest week's data is imported.  If cfg$scope is "EW" the previous
week's data is imported. A print statement gives the week for this and
the Scotland table is updated in the config in global scope
Otherwise a message is printed to the console and no data is imported.
}
\details{
If data is imported the database, table name and variables defined
in the config are used.
}
