% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_weekly_interpol_pops}
\alias{import_weekly_interpol_pops}
\title{Imports interpolated weekly population}
\usage{
import_weekly_interpol_pops(cfg_list)
}
\arguments{
\item{cfg_list}{list containing following fields:
\itemize{
\item{"server"}{SQL server name as a string}
\item{"db_excess"}{SQL reference database name as a string}
\item{"tbl_weekly_deaths_pops"}{Interpolated weekly population table name as
a string}
\item{"var_weekly_deaths_pops"}{Interpolated weekly population variables as a
string}
}}
}
\value{
Interpolated weekly population as data frame.
}
\description{
Imports interpolated weekly population
}
