% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_age}
\alias{is_valid_age}
\title{Data validation: age column}
\usage{
is_valid_age(x, max = 130)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is a numeric value or is NULL/NA, input is between 0
and 130 (for min and max valid ages).
}
