% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_agegrp}
\alias{is_valid_agegrp}
\title{Data validation: age group column}
\usage{
is_valid_agegrp(x)
}
\arguments{
\item{x}{Vector to validate.}

\item{wide}{wide TRUE if the column in wide format (large age bands), default
is FALSE.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is present in the available levels for the type of
agegroups in the column.
}
