% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_agegrp_wide}
\alias{is_valid_agegrp_wide}
\title{Data validation: age group wide column}
\usage{
is_valid_agegrp_wide(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is present in the available levels for the type of
agegroups in the column.
}
