% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_certtype}
\alias{is_valid_certtype}
\title{Data validation: certtype column}
\usage{
is_valid_certtype(x)
}
\arguments{
\item{x}{Check input is numeric between 1 to 14 or NA.}
}
\value{
Vector to validate.
}
\description{
Data validation: certtype column
}
