% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_cestrss}
\alias{is_valid_cestrss}
\title{Data validation: cestrss column}
\usage{
is_valid_cestrss(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check string is either "H", "E", a 5-digit number, or NULL/NA.
}
\details{
Not using this test for now, no unit test yet.
}
