% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_char_length}
\alias{is_valid_char_length}
\title{Data validation: any string column of specific length}
\usage{
is_valid_char_length(x, len = 7)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check string is of a specified length (default 7) or is NULL/NA.
}
