% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_country}
\alias{is_valid_country}
\title{Data validation: country column (ctrynm)}
\usage{
is_valid_country(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is either "England", "Wales", or NULL/NA.
}
