% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_ctry_uk}
\alias{is_valid_ctry_uk}
\title{Data validation: country (UK level)}
\usage{
is_valid_ctry_uk(x)
}
\arguments{
\item{x}{Check input is one of the correct country strings
("England, Wales and non-residents", "England", "Wales", "Northern Ireland",
"Scotland", "Wales"), not NA.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Data validation: country (UK level)
}
