% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_date_format}
\alias{is_valid_date_format}
\title{Data validation: date columns (e.g. dod and dor)}
\usage{
is_valid_date_format(x, min = 19000000, max = format(Sys.Date(), "\%Y\%m\%d"))
}
\arguments{
\item{x}{Vector to validate.}

\item{min}{What is the minimum potential date the data can be, default
19000000.}

\item{max}{What is the maximum potential date the data can be, default is
the system time (run date).}
}
\value{
Logical vector.
}
\description{
Check if either string or integer in of length 8
YYYYMMDD format or is NULL/NA.
}
