% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_dor_week}
\alias{is_valid_dor_week}
\title{Data validation: week number}
\usage{
is_valid_dor_week(x)
}
\arguments{
\item{x}{Check input is positive numeric integer of range 1 to 53, not NA.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Data validation: week number
}
