% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_icd_format}
\alias{is_valid_icd_format}
\title{Data validation: ICD formatted columns (cause of death)}
\usage{
is_valid_icd_format(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check string starts with a letter followed by at least two
numbers, or is NULL/NA.
}
