% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_pod}
\alias{is_valid_pod}
\title{Data validation: place of death}
\usage{
is_valid_pod(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is numeric between 1 and 53, or NULL/NA.
}
