% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_pop}
\alias{is_valid_pop}
\title{Data validation: population}
\usage{
is_valid_pop(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is positive and numeric, not NA.
}
