% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_region}
\alias{is_valid_region}
\title{Data validation: region column}
\usage{
is_valid_region(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is one of the regions or NULL/NA.
}
