% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_sex}
\alias{is_valid_sex}
\title{Data validation: sex and nhsind columns}
\usage{
is_valid_sex(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is either 1, 2, or NULL/NA.
}
