% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_sex_str}
\alias{is_valid_sex_str}
\title{Data validation: sex as string}
\usage{
is_valid_sex_str(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is either "Female" or "Male", not NA.
}
