% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_week_num}
\alias{is_valid_week_num}
\title{Data validation: week number in the year}
\usage{
is_valid_week_num(x)
}
\arguments{
\item{x}{Vector to validate.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Check input is numeric between -53 and 53, or NULL/NA.
}
