% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{is_valid_year}
\alias{is_valid_year}
\title{Data validation: year}
\usage{
is_valid_year(x)
}
\arguments{
\item{x}{Check input is positive numeric integer of length 4, and not
exceeding the current year, not NA.}
}
\value{
TRUE or FALSE depending if the data is in valid format.
}
\description{
Data validation: year
}
