% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{population_processing_totals}
\alias{population_processing_totals}
\title{Calculate population totals by country, age groups and sex}
\usage{
population_processing_totals(weekly_population, current_year)
}
\arguments{
\item{data}{UK weekly population estimates dataframe.}

\item{current_week}{Week for analysis as numeric or string e.g. 1.}
}
\value{
Dataframe with population by \code{week}, \code{country}, \code{agegroup_2} and
\code{sex.}
}
\description{
Takes the processed dataframe containing estimated population in
the UK to totals  by week, country
(including UK and England/Wales/Nonresidents),
age group and sex. Totals are calculated for each variable (except
sum of weeks). The age groups are 5yrs with max "90+".
}
