% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cause_geography.R
\name{prepare_regions_for_analysis}
\alias{prepare_regions_for_analysis}
\title{Derive the different regional breakdowns}
\usage{
prepare_regions_for_analysis(data)
}
\arguments{
\item{data}{Dataframe being used for analysis}
}
\value{
Dataframe containing additional regional breakdowns
}
\description{
From the regional and country columns, create the various levels
of granularity by combining country and regional into a singular region
column. For instance, merging the regions of England, into a single England
region and merging of the England, Wales, and Non-residents, into a region
of the same name.
Nonresidents are only included in the overall region of 'England, Wales and
Non-residents', but not England, Wales or individual English regions, and
don't have  their own separate regional breakdown.
}
