% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_dths.R
\name{process_5y_average_deaths}
\alias{process_5y_average_deaths}
\title{Process death data for the 5 year averages}
\usage{
process_5y_average_deaths(
  finalised_years_deaths,
  prov_years_deaths,
  prov_years
)
}
\arguments{
\item{finalised_years_deaths}{data frame of finalised years deaths data.}

\item{prov_years_deaths}{data from of provisional years deaths data.}

\item{prov_years}{vector of years that provisional data is used.}
}
\value{
data frame of combined finalised and provisional deaths.
}
\description{
Filters the provisional death data for the given year(s). Binds
the data with the finalised data keeping only the columns that are common to
both. Returns a dataframe of the combined datato be used in calculating 5
year averages.
}
\details{
Finalised_years_deaths will already be filtered to required years.
Provisional_years_deaths should only contain the current and previous years.
\code{ctryr} columnn to be dropped from finalised years deaths. \code{dod},
\code{agegrp_5yr}, \code{dor_week} and \code{region} columns to be dropped from
prov_years_deaths.
}
