% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_agegrp}
\alias{process_agegrp}
\title{Process age groups}
\usage{
process_agegrp(data, column)
}
\arguments{
\item{data}{Data frame of unprocessed deaths data.}

\item{column}{Unquoted name of column to process.}
}
\value{
Data frame with extra column.
}
\description{
Re-codes a column with publishing ready age group values.
}
\details{
Changes "+" to "and over", "-" to " to ", "<1" to "Under 1" and
removes any leading 0s.
}
