% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_ew_annual}
\alias{process_ew_annual}
\title{Process deaths for England and Wales from annual file}
\usage{
process_ew_annual(data, nspl)
}
\arguments{
\item{data}{data frame containing annual data for deaths registered in
England and Wales, compatibile with data from 2001}

\item{nspl}{data frame containing pcd column with postcodes and gor column
with codes for regions}
}
\value{
Processed EW weekly deaths data frame - similar to one produced using
provisional data files.
}
\description{
Process deaths for England and Wales from annual file
}
