% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_ew_weekly_deaths}
\alias{process_ew_weekly_deaths}
\title{Process weekly deaths for England and Wales}
\usage{
process_ew_weekly_deaths(data, nspl, max_reg_date)
}
\arguments{
\item{data}{Unprocessed EW weekly deaths dataframe.}

\item{nspl}{NSPL lookup dataframe.}

\item{max_reg_date}{Date of cut off for the week's registrations, as
character.}
}
\value{
Processed EW weekly deaths dataframe.
}
\description{
Re-codes age group, sex, place of death, region
and country columns. Some columns have also been renamed. This data has
validated previously so the values will be as expected. Drops records
with \code{dor} before the max date of registration.
}
\details{
Age groups are processed to publishing format. Sex is re-coded
from 1 to Male and 2 to Female. Missing country is re-coded to "Nonresident".
"Other communal establishment" in place of death is re-coded to "Elsewhere".
reg_stat_dod is renamed to dod.
}
