% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_deaths_model.R
\name{process_excess_deaths}
\alias{process_excess_deaths}
\title{Process excess deaths}
\usage{
process_excess_deaths(data, year, week)
}
\arguments{
\item{data}{dataframe of excess death data}

\item{year}{numeric in format YYYY}

\item{week}{numeric of current week number}
}
\value{
dataframe of excess data formatted and prepared
}
\description{
Process excess deaths
}
\details{
Selects the columns required, converts area to full name, processes
the age group names, creates the new age bands, sums across the new age
groups and filters for the year and weeks required for the publication.
}
