% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_ni_weekly_deaths}
\alias{process_ni_weekly_deaths}
\title{Process weekly deaths for Northern Ireland}
\usage{
process_ni_weekly_deaths(data, max_reg_date, scope)
}
\arguments{
\item{data}{Unprocessed weekly deaths dataframe for NI.}

\item{max_reg_date}{Date of cut off for the week's registrations, as
character}

\item{scope}{string of "UK" or "EW"}
}
\value{
Processed weekly deaths dataframe for NI.
}
\description{
Re-codes age group and sex columns. Some columns have also been
renamed. Country column added. Drops records with \code{dor} after the max date
of registration.
}
\details{
Age groups are processed to publishing format. Sex is re-coded
from 1 to Male and 2 to Female.
If scope  = "EW" then  the previous week is used for NI data so max_reg_date
is one week before.
}
