% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_prev_years}
\alias{process_prev_years}
\title{Process previous years - update when more functions are added}
\usage{
process_prev_years(data, years)
}
\arguments{
\item{data}{Data frame of finalised deaths data.}

\item{years}{vector of required years}
}
\value{
Processed data frame with additional columns.
}
\description{
Processes previous years of data to compute 5 year averages.
}
\details{
Adds week number, week ending and year for dor, filters data to
required years for 5 year average, adds column for wide age group, gives sex
column as factor, adds respiratory/covid columns, adds country column as char
and adds place_of_death_other column.
}
