% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_scot_weekly_deaths}
\alias{process_scot_weekly_deaths}
\title{Process weekly deaths for Scotland}
\usage{
process_scot_weekly_deaths(data, config)
}
\arguments{
\item{data}{Unprocessed weekly deaths dataframe for Scotland.}

\item{config}{list containing week (analytical week), year (analytical year)
and scope (string of "UK" or "EW")}
}
\value{
Processed weekly deaths dataframe for Scotland.
}
\description{
Re-codes age group and sex columns. Some columns have also been
renamed. Country column added. Drops records with \code{dor} after the max date
of registration.
}
\details{
Age groups are processed to publishing format. Sex is re-coded
from 1 to Male and 2 to Female.
If scope  = "EW" then the previous week is used for Scotland data so
max_reg_date is one week before
}
