% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_asmr_ctry_sex.R
\name{process_uk_deaths_asmr}
\alias{process_uk_deaths_asmr}
\title{Combines England and Wales, Scotland and Northern Ireland and calculate UK
counts of deaths by country, age groups and sex}
\usage{
process_uk_deaths_asmr(
  ew_df,
  scot_df,
  ni_df,
  current_year,
  current_week,
  scope,
  esp,
  weekly_pop
)
}
\arguments{
\item{ew_df}{Processed England and Wales deaths dataframe.}

\item{scot_df}{Processed Scotland deaths dataframe.}

\item{ni_df}{Processed Northern Ireland dataframe.}

\item{current_year}{Analysis year as numeric in format YYYY.}

\item{current_week}{Week for analysis as numeric or string e.g. 1.}

\item{scope}{String of "UK" or "EW".}

\item{esp}{European Standardised Population (ESP) lookup data frame.}

\item{weekly_pop}{Weekly population interpolation estimates dataframe.}
}
\value{
Dataframe of UK counts of deaths by country, age groups and sex.
}
\description{
Filters out any data outside of the analysis time period and
combines datasets for England and Wales, Scotland and Northern Ireland. Then
counts deaths by week, country, age and sex. If no deaths occur in a group,
counts of 0 are added. If the scope is EW then there are no rows for UK, NI,
Scot for the latest week.
}
\details{
Adds a column of 1 for \code{deaths} for England and Wales as the data is
record-level. Drops Scotland counts for deaths involving Covid to leave "all
deaths" as Covid are duplicated counts.
}
