% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{pub_week_year_validation}
\alias{pub_week_year_validation}
\title{Check the week and year derivative from the data}
\usage{
pub_week_year_validation(data)
}
\arguments{
\item{data}{Processed deaths dataframe.}
}
\value{
List of counts of NA or impossible week numbers or years. Message of
how many dates are unable to be derived.
}
\description{
Uses \code{dor} and \code{reg_stat_dod} derived publication weeks and
years to check if there are any dates that are unable to be derived or
impossible values (registration before occurrence or negative).
A small proportion of deaths are registered on the same day as they occur,
but none should be registered before they occur.
It reports the number of records that could not be assigned a week or year
number.
}
\details{
IDEA: Add a second argument of the quality report to save the output
of this function, rather than just a message.
}
