% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{qa_table_4_counts}
\alias{qa_table_4_counts}
\title{Create QA table comparing counts between table 1 and table 4 for England,
Wales and Non-residents.}
\usage{
qa_table_4_counts(table1, table2)
}
\arguments{
\item{table1}{data frame containing table 1 England and Wales death
registrations data}

\item{table2}{data frame containing table 4 England and Wales death
registration delays data}
}
\value{
a QA table comparing table 1 and table 4 counts for England,
Wales and Non-residents.
}
\description{
Create QA table comparing counts between table 1 and table 4 for England,
Wales and Non-residents.
}
