% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{qa_table_6_counts}
\alias{qa_table_6_counts}
\title{Create QA table comparing counts between table 1 and table 6 for England,
Wales and Non-residents, England and Wales}
\usage{
qa_table_6_counts(table1, table2)
}
\arguments{
\item{table1}{data frame containing table 1 England and Wales death
registrations data}

\item{table2}{data frame containing table 6 UK death registrations data}
}
\value{
a QA table comparing table 1 and table 6 counts for England,
Wales and Non-residents, England and Wales.
}
\description{
Create QA table comparing counts between table 1 and table 6 for England,
Wales and Non-residents, England and Wales
}
