% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cause_geography.R
\name{registration_cause_zero_values}
\alias{registration_cause_zero_values}
\title{Identifies missing values in by cause of death columns and replaces with
0 counts}
\usage{
registration_cause_zero_values(data)
}
\arguments{
\item{data}{Data frame with week, week ending, region, cause of death,
counts}
}
\value{
Dataframe containing death nums for causes, by region, per week,
including 0 counts
}
\description{
Assesses the England, Wales, and non-residents cause of death
data frame to identify and replace date columns where there are no deaths
recorded for that week by certain causes
}
