% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{reverse_max_agegrp}
\alias{reverse_max_agegrp}
\title{Reverse the maximum aggregations of the 5-year age bands to match
European Standard Population (ESP) age bands}
\usage{
reverse_max_agegrp(data)
}
\arguments{
\item{data}{Dataframe with processed \code{agegrp_5yr} column.}
}
\value{
Dataframe with \code{agegroup_2} which maxes out at \verb{90 and over}.
}
\description{
ESP top age group is "90 and over" so we must align for excess
deaths. Also renames \code{agegrp_5yr} to \code{agegroup_2}.
}
