% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round}
\alias{round}
\title{Round 0.5 up}
\usage{
round(x, digits = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits to round to, default 1}
}
\description{
Replaces R round function with janitor::round function across the
package. Not currently exported.
}
