% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{subtotals_check}
\alias{subtotals_check}
\title{subtotals_check}
\usage{
subtotals_check(data, data_type, topic)
}
\arguments{
\item{data}{The R dataframe in the global environment containing the
relevant data.}

\item{data_type}{string identifying the type of data ("registrations",
"occurrences")}

\item{topic}{string of grouping of interest (age, sex, region, place).}
}
\value{
TRUE or FALSE if table exists and subtotals sum correctly.
"Check table exists" if table doesn't exist.
Nothing will be saved to the global environment by running the function
alone.
}
\description{
Checks whether tables exists, calculates information for doing subgroups
checks and checks that the subtotals sum to the correct overall value
}
\examples{
x <- subtotals_check(ew_registrations, "registrations", "age")

}
