% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_dths_ctry_age_sex.R
\name{total_by_group}
\alias{total_by_group}
\title{Calculates totals based on given groups (inc. year, week)}
\usage{
total_by_group(
  data,
  grouping_cols,
  total_col,
  total_name,
  sum_colname = "deaths"
)
}
\arguments{
\item{data}{Long format dataframe with \code{year} and \code{week}, plus any
columns to group by and total to create for.}

\item{grouping_cols}{Vector of strings for column names to group by.
Note: \code{year} and \code{week} are always grouping variables so do
not need to be included in this vector.}

\item{total_col}{String of column name to total by.}

\item{total_name}{String of what to populate the total column with.}

\item{sum_colname}{String of the column to sum, default is "deaths", other
option could be "population".}
}
\value{
Dataframe of the total counts of deaths.
}
\description{
Takes an input of chosen grouping variables (plus \code{year} and
\code{week}) and then creates rows for the totals of each group combination.
Defines the column to total by and what should the total rows be "called".
Optional argument can be given to define the numeric column column name to
sum by, default is \code{deaths}.
}
\details{
Column order of returned dataframe is dependent on grouping_cols and
total_col arguments.
}
