% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{validate_data}
\alias{validate_data}
\title{Validates deaths weekly EW data}
\usage{
validate_data(data)
}
\arguments{
\item{data}{Unprocessed deaths dataframe from SQL, weekly deaths.}
}
\value{
Report list object containing each of the checks by counts and
indices of the passing, failing and nulls in each one. Returns warning if
there are any failing records.
}
\description{
Uses subfunctions depending on the column of data frame to
validate if the specific column content is expected format.
The results of the validation are compiled into a report, displaying the
count and indices of pass, fail, and null entries for each column.
}
\details{
All subfunctions allow for NA, plus specific validations.
}
