% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_and_checks.R
\name{validate_weekly_pop}
\alias{validate_weekly_pop}
\title{Validates weekly interpolated population}
\usage{
validate_weekly_pop(data)
}
\arguments{
\item{data}{Unprocessed weekly interpolated dataframe from SQL (output of the
excess deaths pipeline).}
}
\value{
Report list object containing each of the checks by counts and
indices of the passing, failing and nulls in each one. Returns warning if
there are any failing records.
}
\description{
Uses subfunctions depending on the column of data frame to
validate if the specific column content is expected format.
The results of the validation are compiled into a report, displaying the
count and indices of pass, fail, and null entries for each column.

@details Specifically fails (errors) if there are any NA in the whole
dataframe.
}
