% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_age_sex_geo_place.R
\name{week_ending_from_week}
\alias{week_ending_from_week}
\title{Uses existing combinations of week and week ending to calculate missing week
ending dates}
\usage{
week_ending_from_week(data)
}
\arguments{
\item{data}{Processed deaths counts dataframe with week and week_ending
columns.}
}
\value{
Original dataframe with week_ending column recreated and added to
end.
}
\description{
Due to adding counts of 0, some week_ending are NA and must be
re-populated with week_ending derived from the week number.
}
